backend/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
httpx>=0.27.0,<1.0.0
numpy>=1.26.0,<2.0.0
Pillow>=10.4.0,<13.0.0
pandas>=2.1.0,<3.0.0
dashscope>=1.15.0
open-clip-torch>=2.24.0
umap-learn>=0.5.6
hdbscan>=0.8.33
segment-anything>=1.0
transformers>=4.40.0
torch>=2.1.0
